requirements.txt
aws-cdk.aws-codebuild==1.15.0
aws-cdk.aws-codepipeline-actions==1.15.0
aws-cdk.aws-codepipeline==1.15.0
aws-cdk.aws-events-targets==1.15.0
aws-cdk.aws-events==1.15.0
aws-cdk.aws-lambda==1.15.0
aws-cdk.aws-s3==1.15.0
aws-cdk.aws-ssm==1.15.0
aws-cdk.aws-stepfunctions-tasks==1.15.0
aws-cdk.aws-stepfunctions==1.15.0
aws-cdk.core==1.15.0

black
pylint
rope
